Starting work on the endgame table generator.
Calculate table size as a BigNumber, so I don't have to
worry about integer overflows going undetected.
When I cast (int)BigNumber, it will throw an exception
if the value does not fit in an int.

Write tables to disk containing all zero bytes.
Next I will start filling in the generator logic.

+++ src/EndgameTableGen/TablePrinter.cs
@@ -1,11 +1,12 @@
 using System;
+using System.Numerics;
 
 namespace EndgameTableGen
 {
     internal class TablePrinter : TableWorker
     {
         private int tableCount;
-        private long totalSize;
+        private BigInteger totalSize;
 
         public override void Start()
         {
@@ -21,7 +22,7 @@ namespace EndgameTableGen
             for (int m=0; m < WorkPlanner.NumNonKings; ++m)
                 Console.Write(" {0}{1}", config[0,m], config[1,m]);
 
-            long size = TableSize(config);
+            BigInteger size = TableSize(config);
             Console.WriteLine(" {0,20:n0}", size);
 
             totalSize += size;
